# swagger/spring-server/src/main/java/io/lozzikit/survey/api/endpoints/SurveyApiController.java
package io.lozzikit.survey.api.endpoints;

import io.lozzikit.survey.api.SurveyApi;
import io.lozzikit.survey.api.model.Survey;
import io.lozzikit.survey.services.SurveyService;
import io.swagger.annotations.ApiParam;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.servlet.support.ServletUriComponentsBuilder;

import java.net.URI;

@Controller
public class SurveyApiController implements SurveyApi {
    @Autowired
    SurveyService surveyService;

    @Override
    public ResponseEntity<Void> addSurvey(@RequestBody Survey body) {
        String newSurveyId = surveyService.saveSurvey(body);

        URI location = ServletUriComponentsBuilder
                .fromCurrentRequest().path("/{id}")
                .buildAndExpand(newSurveyId).toUri();

        return ResponseEntity.created(location).build();
    }

    @Override
    public ResponseEntity<Survey> getSurveyById(@ApiParam(value = "ID of survey to return", required = true) @PathVariable("surveyId") String surveyId) {
        Survey survey = surveyService.getSurvey(surveyId);

        if (survey == null) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        } else {
            return new ResponseEntity<>(survey, HttpStatus.OK);
        }
    }
}