# swagger/spring-server/src/main/resources/api-spec.yaml
---
swagger: "2.0"
info:
  description: "HEIG-VD AMT course Survey microservice\n"
  version: "1.0.0"
  title: "Survey microservice"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/proteck/AMT/1.0.0"
tags:
- name: "survey"
  description: "Manage some surveys"
  externalDocs:
    description: "Find out more"
    url: "http://swagger.io"
schemes:
- "https"
- "http"
paths:
  /surveys:
    get:
      tags:
      - "survey"
      summary: "Get all surveys"
      operationId: "getSurveys"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Survey"
      security:
      - api_key: []
  /survey:
    post:
      tags:
      - "survey"
      summary: "Add a new survey to your account"
      operationId: "addSurvey"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Survey"
        required: true
        schema:
          $ref: "#/definitions/Survey"
      responses:
        201:
          description: "Return the id of the created survey"
        405:
          description: "Invalid input"
      security:
      - api_key: []
  /survey/{surveyId}:
    get:
      tags:
      - "survey"
      summary: "Find a survey by ID"
      description: "Returns a single survey"
      operationId: "getSurveyById"
      produces:
      - "application/json"
      parameters:
      - name: "surveyId"
        in: "path"
        description: "ID of survey to return"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Survey"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Survey not found"
      security:
      - api_key: []
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Survey:
    type: "object"
    required:
    - "owner"
    properties:
      owner:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      description:
        type: "string"
      createdAt:
        type: "string"
        format: "date-time"
      questions:
        type: "array"
        items:
          $ref: "#/definitions/Question"
    xml:
      name: "Survey"
  Question:
    type: "object"
    required:
    - "question"
    properties:
      question:
        type: "string"
    xml:
      name: "Question"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"